backend/pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "videonotes-backend"
version = "0.1.0"
description = "FastAPI backend for VideoNotes with YouTube integration and AI analysis"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "VideoNotes Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "pydantic-ai-slim[openai,groq]>=0.1.0",
    "yt-dlp>=2023.9.24",
    "youtube-transcript-api>=0.6.1",
    "asyncpg>=0.28.0",
    "pgvector>=0.2.3",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.24.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.0.290",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W"]
